requirements.txt
torch>=2.4.0
torchvision>=0.19.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
huggingface-hub>=0.23.0
safetensors>=0.4.3
jsonschema>=4.22.0
peft>=0.7.0
paddleocr>=2.7.0 
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddlepaddle-gpu>=2.5.0  # GPU version
scikit-image>=0.21.0
python-pptx>=0.6.21 
pywin32>=305  
pdf2image>=1.16.0

# --- CUDA-only (skip on Windows) ---
# xformers
# bitsandbytes>=0.41.0